Button step change with modifier key
Most plus/minus buttons now have 2 different step sizes in each
direction. The normal one (standard mouse click on plus/minus) or the
modified one (Ctrl pressed and mouse click).
The step differences depend on the function, for some (e.g. work width)
it's standard 0.1 and modified 0.5, for others it's 1 and 5, or 5 and
10). The modifier key is the one used in InputBindings, so it can be
changed by the user).