Update generateCourse.lua
So buddys I have come to a quite  good version after going through the offset routine many times.  The offset can now be done util ther is nithing more to offset, have too trie it a bit in complex cases. Overlapping of the last row can be set by the vehicle.cp.headland.overlap constant ( set to 1/4) coud be added to the menu to be changed from -1 to 1/2. I let you try it and tell if somthing is wrong. now have to fine up the transition between the headlanes and add a bit smoothing for long segments .