Update mode9: fully unloading the shovel
When bunker is full while unloading (timescale 1x), the unloading isn't continous and will unload only small portions. Those portions are less than 1% and so the shovel won't unload completely with the statement 
if fillLevelPct <= 1 or stopUnloading then

In a first test with the liebherr 432 with the standard multi shovel the new line (fillLevelPct == 0) worked well.