Enhanced Speed Control
This provides automatic speed control while not on fields.  It uses the difference of the upcoming way point angles to adjust the current speed.  When selected it will be in control when the "from recording" or "maximum speed" setting would normally be used.

On the Speed limits tab the "Speed" setting has a new option titled "enhanced speed control".  When selected two new settings are available.

The "Look ahead" setting determines how far ahead on the course to look for turns.  If the vehicle takes a long time to slow down this value will need to be increased.  This will cause turns to be found earlier, allowing the vehicle to slow down earlier.

The "Angle ranges" and "Target speeds" paired values work together to determine the speed based on the waypoint angle differences.  There are up to 9 pairs of the angle and speed values.  If the angle difference of the waypoint is less than or equal to the angle specified on this setting the corresponding interpolated speed will be used.  If the angle difference is greater than the last angle in the list it will not go below the last speed in the list.  If the last angle listed is less than 90 degrees then another value pair will appear up to 9 total.

This essentially allows you to specify a speed curve based on the angle differences.  An example is below.

Angle ranges:   0.0   1.5   3.0   5.0  10.0  90.0
Target speeds:   50    40    35    30    10    10

If the current waypoint and the next waypoint have the following angle differences the corresponding speed would be used for that waypoint.  Note that the look ahead setting will be used to find waypoints further down the course.  If the waypoint speed is less than the current speed then it will be averaged into the final speed calculation.

Angle Difference	Vehicle Speed
  0.0 (straight route)	50
  1.0			43
  1.5			40
  2.8			36
  3.0			35
  8.0			18
 10.0			10
 45.0			10
110.0			10