Shovelmode 9 improvement
I had some problems with it. First of all I didn't even know, that CP is
trying to empty the entire Silo by creating a map and divide it in small
fillUnits.
In front of the Silo on the right wall there was about 40l silage. CP
wanted to load this and was driving allways to the end of the silo. I
couldn't get it to start shovel the rest of the silo.
This is why I changed the fillUnitLevel check >0 to >200 or 10% of
shovelVolume
The Second problem I had. There were 500l Chaff in the wall. CP always
wanted to load it. But because it was completely in the wall it was
impossible to shovel anymore in the silo.
Therefore I let the shoveler always look for a new fillUnit when it
drives into the silo. I also try to compare it with the fillType which
is mostly in the silo.

I created a Video to show the changes of the shoveling in action:
https://youtu.be/gus4_NtwZLo