Self Unloading Chaff Havesters
Added need unloading to combines. Not sure if it was missing or is a bug of my update. Know bugs that are related, pipestate is related to something in unload tipper which toggles it but does inhibit functionality, it is also related to turn on but this is good behavior. Bug not related but need testing for confirmation. If a chaff harvester sits to long when finish work he tries to start again. Turn off field harvester is not square to the field when finish though this is not related as the beet harvester does this as well