Minor Tweaks
Increased Turning. Radius 14m was working but steering had to be turned to the max to make the turn. 18m works better steering no longer is maxed out to turn. Also the offest probably is suppose to be .4m has it leaving a line of unplough set it to .4 to fix this.