v01.00130 Final Final Run Counter Fixes
Fix Kotte Universal mod issue. Prohibits changing of total number of runs if runs have been completed. Also prohibits decreasing beyond 1 the difference between number of runs and total number of runs if in the middle of loading. This was all done to prevent weird behavior if the total number of runs was changed during loading.