v02.00005 Align with first waypoint of course
When driving to the first waypoint, an alignment course is added
which makes sure the vehicle arrives at the first wp in the
correct direction.

Works in mode 2 when trailers full and in mode 4 and 6 when heading
for the first waypoint of the fieldwork course.

Also, in dev mode allows for recording a course with the first
three waypoints much closer for tight fields.

- generate more waypoints on the circle to the first
- swap in a temporary course for the alignment (not just one WP)
- slow down before the alignment turn