5.02.00091 Fixes #2717
Ridge marker calculation fixed. Did not work correctly when
the course gen used multiple blocks.

Won't add ridge markers to the first and last row of the block
to avoid conflicts with the adjacent blocks.

Still don't know why are the ridge markers used to decide where
the fruit is in mode2, instead of actually checking the fruit.