v05.03.00044 MR and Real Fill Type Mass Mod Fix
This fixes the problems with MR and RFTM driving like a running away frieght train. Turns must still be taken easy has CP only knows the recored speed and will try with brakes to keep it there but you can still exceed the traction force of the ground cause CP does not know if it needs to slow down more than recored because of a sharp turn. May or may not have a bug were it will refuse to swicth direction and cause a slipping message. Unkown cause at this time and reproduction unkown