Use smaller grid but more tiles
This solves the problem of a waypoint near the boundary of a tile:
if only that tile is reserved, a vehicle in an adjacent tile still may
collide.

Therefore, use smaller tiles and always reserve the adjacent tiles as well.
This results in more tiles to keep in the reservation matrix but as our
lookahead time is still short this should not be an issue.