Fixe for lowering tools(nil error)
Added a check to ensure jointDesc was not nil before check if the tool can be lowered. Also updated the FillUtil.FILLTYPE_UNKNOWN in the last push to the correct 19 version. Also updated mode4 for lowering tools. Appears to work tho there was another error with the tool being used to prevent a real check