Fix for loading of trains
The loadFinished() function returns a value, this has to be passed on when appending. Appending does not
actually do this so  you always need to overwrite such functions, call the super func, catch the return
value, and then return that yourself. Now trains successfully load from placeables (the vehicles got
loaded synchronously there so only were an issue on placeables. Everywhere else they are loaded async.)