Realistic wages in CpManager.lua
I know, this is a very minor thing right now, but I'd at least like you to consider this change. I know of no farm help (in the United States) that makes $1,500.00 per hour. I would assume it's the same whether we're talking dollars, pounds, euros, or whatever currency.

I've always made this change since FS 2013. In FS17 I used 12.50 as the wage value, and the way the wage is figured, it actually became $16.75. Some maps would even boost it somehow to 20 - 22 per hour which is still pretty high, but a far better figure from a realism standpoint. Thank you for your time, and keep up the great work!