v6.01.00057 traffic collision fix for #3187
The CollisionDetector was created in the AIDriver constructor but
seems that at that time getAttachedImplements() did not have any of
the attached tools so we did not put those on the ignore list and ended
up after start with our weight as a colliding object. Should look into
the startup sequence as the AIDriver is constructed twice...

Now creating the CollisionDetector on demand, always removing it before
start so it always has the latest state.