v6.01.00068 Fieldwork offsets improved #3134
This should fix a number of issues when transitioning from the
headland to the up/down rows (not the turn maneuvers though)

As the Course class has the offsets now it is not needed to adjust
temporary courses for the offset at all.

Also, in mode 4 and 6 offset is applied to the fieldwork part only.