v6.01.00099 Use the Giants AI events only to unfold raise/lower
Looks like it is safe to rely on the Giants onAIImplementStart and
onAIImpleentEnd logic to raise/lower and unfold implements.

onAIImplementEnd by design won't fold them so that is the only case
we have to explicitly unfold them when leaving the fieldwork for a
refill/unload course.

For the future we should use this AI for all implements and special tools
(like with the bale loader) and implement/change the AIImplement event
handlers only, so the fieldwork AI driver code does not need to be changed
for any tool specific behavior.