6.01.00195 Mode 4 and 6 pathfinding changes
- after unload/refill, return to a previous waypoint to avoid leaving
unworked areas (#3626)

- use pathfinding when starting the field work and when returning to
the first point (this helps #3535). For now, always return to the first
point, we'll have to add this option to the HUD later and remove from
the course generator.

- when returning to the first point, raise and fold implements (#3583)
Note that this does not work if your course is generated with the 'return
to first point' option, see above.