6.01.00142 Course gen: remove empty blocks #3423
On non-convex fields opposite headlands may overlap and
then the course gen tries to generate up/down tracks in
the overlap. There won't be any tracks generated but a
block with zero waypoints remains there and so the last
waypoint will be a turn.

Make sure such empty blocks are removed.