6.01.00144 headland turn threshold increased fixes #3424
When a course with headland turns started in a field corner the turns
will be well over 90 degrees and the turn system uses a normal up/down
row change turn. This can be a problem when symmetric lane change is
activated as it switches the left/right sides on the headland where it
should not.

Increased the angle threshold so now everything under 150 degrees will
perform a headland turn.