6.01.00163 Turn maneuver tuning #3485
When a wheeled tool does not have a decent front/back marker or the
distance between the back marker and the tool's turning node is bigger
than the reverse waypoint switching distance (3m) then the first,
reversing waypoint will be in front of the tool and it'll jackknife
while trying to reverse to it.

Make sure that if the next turn waypoint is reverse but it is in front
of the tool we switch to the next waypoint. This is a lot easier than
fine tune every individual turn maneuver to make sure the reversing wp
is always behind the tool.

(Of course this would not happen with PPC driving but refactoring
turn.lua to make that happen will have to wait now)