6.01.00182 Fix sprayers leaving unworked patches at end of row
#3594 and #3242

Sprayers have a rectangular work area but the spray is at an angle so
the front of the area is not covered.

Also, some sprayers have multiple work areas depending on the configuration
and fill type which can also move the front or back markers.

This leads to little unsprayed rectangles at the row ends as the turn
code turns off the sprayer when the back marker (which is actually in the
front when the implement is attached to the back of the vehicle) reaches the
field edge.

So, if both the front and back markers are from sprayer work areas, move
the back marker to where the front marker is which will result turning
off the sprayer later.