6.01.00210 Fixes #3748
Reverting #3554 again. That change is not bad but looks like turn.lua
has a bug with stage 4, the current waypoint remains at the last reversing
one but driveToPoint() still drive forward with a dz -1 so covers the issue.

This will need some refactoring of the turn code to fix properly.