6.01.00211 Proper fix for #3748, driveInDirection enabled again
The last forward section of the tractor headland turn started back
at the implement, behind the tractor, so we used a dz < 0 to reach
the turn end. This for whatever reason worked with driveToPoint and
the tractor still drove forward, but driveInDirection (correctly) turned
back.

Do not a add an entire section now, just one waypoint well in front of
the tractor at the end of the turn to make sure stage 3/4 finishes and the
point is in front of the tractor.