6.01.00218 Fix alignment course generation #3800
The straight sections generated by the turn code had the last waypoint
with the same coordinates double. Then, the alignment generation appended
a circle with a starting point with the same coordinates. This resulted
in having the same waypoint three times in the course, messing up the
angle calculation between the waypoints.

Made sure that generateTurnStraightPoints() does not add a second end
waypoint and when called from the align wp generation, does not add a
last waypoint at all.