6.01.00225 Driving with offset improved #3285 #3667
The waypoint's direction (Y rotation) was calculated at course
initialization with no offset. This resulted in wrong wp directions at
high offsets in corners, confusing the PPC and the vehicle wandering
off course.

Calculate the waypoint's Y rotation on the fly instead of using the
value calculated at the course initialization. This makes sure the
direction is always correct and pointing to the next waypoint.

This still results in weird behavior in corners with very high offsets
but vehicles should now stay on the course.