6.01.00230 turn maneuver refactoring
Create a front and a back marker node at the front/rear of the work area
and use these to determine when to lower/raise implements.

Also, let sowing machines stop while lowering to make it more realistic

Combine headland turn fixed, it lowered the header at the turn start
waypoint, but in headland turns lowering is controlled by the turn wp
attributes.

Combine wide turns made tighter to stay as much on the field as possible.

Combine headland turn rework according to #3760

Make pocket in corner only when turn on field is enabled
Use regular turn maneuvers in 270 corners

Using a new turn method, no turn.lua involved in these now, only
in headland turns for combines.

Detect combines slipping and try switching to the next waypoint when
it does. This an experiment to handle headland corners on very tight
fields where the combine may get stuck in fences or trees very close
to the field edge.