6.01.00232 Use the front/back marker nodes in turn
Use the front/back marker nodes calculated by the AI driver in the
turn.lua code instead of the old aiFrontMarker and backMarkerOffset.

This seems to work better, considering all work areas and now the
cotton harvester is performing nice turns.

If you see issues with this just revert back to the previous release...