6.01.00251 Move grabber to transport position on unload course, fixes #3526
for some bale loaders like the Anderson RBM2000, the onAIImplementEnd()
does not work seemingly because the grabber is still busy
with the last bale when switching to the unload course.

Make sure here that it is moved to transport position on the unload course.

Did not test with other balers, hopefully this fix did not break those.