6.01.00262 Better turns for tractors with implements without wheels
Use a fishtail (or K) turn for anything easy to reverse, that is has non
towed implement, like tractors with implements mounted on the 3 point
hitch as it takes up a lot less space on the headland.

Our calculation of how much space is needed is still off a bit
so you may have to turn off 'turn on field' for this to work for tractors.