6.01.00271 Improve combine turns with offset, #4068
After the turn, find the next waypoint ahead of us and initialize the
PPC to that waypoint instead of the turn end as the turn end may be behind
us if we are on the inside of an offset corner.

Also, fixed the inner headland turn for combines to use the offset.

This release should work fine with multiple tools as long as 'turn on field'
is activated. With multiple tools and 'turn on field' deactivated the
inner combine will still not turn nice, I'll fix that in a later release.