6.01.00283.1 Beta: AutoDrive destinations as virtual courses
BACKUP YOUR COURSEPLAY COURSES BEFORE TRYING THIS! I messed a lot
with the course management and can't guarantee I did not screw up
something.

When AutoDrive present, add a virtual course folder with AutoDrive
destinations to the course management. When loaded, AD will create a
course from the current vehicle position to the selected destination.
(You'll need AutoDrive 1.0.4.9 or later for this to work correctly)