6.01.00280 Wide turn fix for implements front and back #4082
The wide turn calculation is a little weird as it is using a
reverseOffset even if there's no reversing involved to calculate the
circles and that calculation is different for implements on the front
of the vehicle.

Changed reverseOffset calculation so that branch is called only when
there are no implements on the back.