6.01.00283.3 Beta: AutoDrive load and speed fixes.
BACKUP YOUR COURSEPLAY COURSES BEFORE TRYING THIS! I messed a lot
with the course management and can't guarantee I did not screw up
something.

Have some primitive speed control for corners if there is no recorded
speed. (#3962, #4108, #4110)

After startup we are loading the courses into the vehicles but AD does
not seem to be present at that time yet, so skip loading if this is
the case #4114.