6.01.00286 Load courses at startup only if configured
To speed up loading savegames, closes #4087

Courses not assigned to vehicles are only loaded when configured in
savegameXX/courseplaySettings.xml, like this:
<loadCoursesAtStartup active="true"/>

By default, this is set to false, so CP won't load a course unless a
vehicle has that course loaded. If you have problems like described in
#4016, you need to save your game and then edit courseplaySettins.xml
to set loadCoursesAtStartup to true.