6.01.00294 Don't use calculated radius in fieldwork.
Fieldwork uses the radius supplied by the course generator on island
bypass. When we calculated the turn radius on the fly for AutoDrive
courses that also added a radius to fieldwork courses triggering the
tight turn offset mechanism causing all kinds of issues like #4172, #4173
and possibly #4156.

Separate the two radii and use only the course generator supplied ones
during fieldwork.

Also, fieldwork courses generated since 287 should now be regenerated as
when they are saved and loaded they'll contain the radius.