6.01.00311 Headland turns fixes #4391 and possibly #4372
Fixed calculation of implement raise point on headland turns.

Added option for implement raise/lower times (early/late) to the
advanced settings.

Raise early: raise implement when the front (AI front markers) of the
implement reaches the field edge or headland edge.

Raise late: raise implement when the back (AI back marker) of the
implement reaches the field edge or headland edge.

Lower early: lower implement when the front (AI front marker) of the
implement reaches the field edge or row start.

Lower late: lower implement when the back (AI back marker) of the
implement reaches the field edge or row start.

Also fixed loading of boolean setting: don't change setting if it is not
present in the save file.