6.01.00358.1 Beta Use pathfinding in wide turns
May help with #4091 and #4763 and similar.

This is a proof of concept to use the new hybrid pathfinding for
wide turns which should keep the vehicle off the fruit and on
the field during the turn.

You can most likely see it working when using the racetrack pattern
or skipping rows and turn on field is enabled.

WARNING: This version includes a lot of refactored code and may be even
buggier than what you are used to. I have done very limited testing on
this as I don't have the time to try every combination of vehicle/
implement and all shapes of fields and courses.