6.01.00359 Click on a vehicle to switch to it (#4784)
Another great new feature from Russ B:

* New feature: Click to switch vehicles

This was developed as a way to quickly jump from your current vehicle to another vehicle by mouse clicking on the vehicle on-screen.
This can provide a quicker way of getting into a vehicle which you can see from your current position rather that cycling through your vehicles with a hotkey.  You'll need to enable this in the hud advanced settings page.

NOTES
1) The vehicle must be visible and within range (1000 meters, which is pretty well a pixel on the horizon) to be able to switch to it.
2) You must click on the power unit.  Ex. For a tractor with a trailer, you must click on the tractor.  Clicking on the trailer won't do anything.
3) The only input used is the 'Courseplay: mouse action' hotkey.  No other configuration or hotkeys are required.