Fixed AIDriver.lua:715: attempt to index field 'course' error
Error encountered on 00359Dev while starting to drive course in Mode 2 if the driver has not driven a course yet this session.

To duplicate this issue:
1) With a truck, create a simple straight course as shown and save it.
2) Switch to mode 1 (grain transport)
3) position the about a third of the way up the course (although the position doesn't matter)
4) save the game.
5) reload the game and switch to the truck used to create the course earlier.  You should see it on mode 1, 1/3 of the way along the course.
6) set it to mode 2
7) drive the course
8) the hud disappears and the log starts filling with:
AIDriver.lua:715: attempt to index field 'course' (a nil value)
9) The truck continues to drive and other vehicles seem to run their courses ok, but the truck's hud can't be brought back up and the errors continue to fill the log. The only solution is to restart the game.
10) Note: If at step 6 above you started to drive while on mode 1, then stop driving, then switch to mode 2, then start driving, the errors do not occur.

SOLUTION
The driver.course is nil when starting to drive a mode 2 if the driver has not yet driven a course on a different mode.  This originates in hud.lua calling getCanShowDriveOnButton.

In AiDriver.lua, check to see of course is nil.

function AIDriver:isNearFillPoint()
	if self.course == nil then
		return false
	else
		return self.course:havePhysicallyPassedWaypoint(self:getDirectionNode(),#self.course.waypoints) and self.ppc:getCurrentWaypointIx() <= 3;
	end
end

now when you do step 7 above, the error does not occur and the truck drives it's mode 2 course.