6.01.00363 Plow and baler fixes
Fixes #4737, partial fix for #4802

Make sure to reset all states before all unloads to prevent
false detection of bales.

Also, partially fix the symmetric lane change in the HUD. That
must be revisited as after the HUD refactoring the addRowButton()
function now can't pass in a parameter to the callback anymore,
instead it seems to use the 'line' instead.

Fixed 180 AI turns with rotating plow to consider reversed
offset after plow is rotated during the turn.