6.01.00372 Fix for #4836
When there are many headlands we set the pathfinding's
range too high so when doing the A* for the middle part
we immediately (and incorrectly) report finding the goal.

This results in a path with a single waypoint (the goal node)
which the code was not able to handle.

For now, just bail out in this case to prevent the crash. Will
fix it properly later.