Call AutoDrive to refuel (#4889)
Only if unload or refill with AD is active.
Also fixed not loading saved setting for use autodrive if it was park or refill and park.
Needs new AD update (already pull requested by me) but keeps compatibility with previous versions, just doesn´t exit fiel with pathfinding if AD not updated