6.01.00385 Check for 'turn on field' in AI Turn, fix for #4908
When AI Turns was on we did not check for the 'turn on field' settings
and vehicles left the field during the turn.

Check if there's enough space on the field to execute a 3 point AI turn,
if not, revert to old turn maneuver even if AI Turn is on.

Also there's slip detection back (no reaction just display the fact)
for #4878