Awesome new feature: save/restore vehicle positions
This is for testing. Use Alt+Space to save the position
of the current vehicle, then drive it wherever you want.

You can save the position of multiple vehicles.

Then use Ctrl+Space (does not always work) to restore the
position, or use the console command cpRestoreVehiclePositions
to restore the position of all vehicles which have a saved
position.