Mode 2 merge, work in progress
Don't use pathfinding in the very last stage of approaching
a stopped combine, just rely on the PPC to find the way to the
pipe.

This should work fine as long as CP drives to the combine as
it will approach it from the correct angle.

If the unloader is manually started it must be in a position where
it can pick up the course under the pipe.