Fixed errors raised when stopping courseplay workers in multiplayer. (#5327)
Fixes a lua crash when a courseplay worker stops in multiplayer. I was hoping for a nicer fix than this, but this seems to resolve the issue without adding too much risk. Hopefully this can be tidied up in the future? (I'm not too sure why AIDriver.stop is called on the client, but start isn't for example). I think this fixes a host of issues:
#5273
#5240
#5223
#5150
#5087
#5061
#5013