6.02.00015 Mode 2: unload moving combines
Unloader will now drive to moving combine and
start unloading if the fill level reaches the limit set in HUD.

Note that it won't drive to combine as long as the pipe is in the
fruit.

Still _no_ support for multiple unloaders for the same
combine/chopper. Expect accidents to happen if there is more than
one unloader and combine/choppper on the same field.