Pathfinder tweaks to improve large field behaviour
The h term in the A* algorithm is scaled too low to really do enough
when working on a path across a large distance. Quick and dirty scale by
50% pushes it more towards the "greedy" end and away from the "dijkstra"
end, though this does mean it likes to stick close to obstacles until it
gets a straighter shot to its goal. Need a better way to estimate that h
term.

Tried to hack together a way to have the pathfinder try something
different if it fails. For mode 2 it initially paths on the field, off
the fruit. If that fails, it tries looking further off the field. If
THAT fails it then tries a straight shot through the fruit as well.