AutoDrive support for level/compact
This will stop an AD vehicle in the search range of the compacter.
As long as the compacter is calling the function courseplayer.spec_autodrive:HoldDriving(courseplayer) the AD vehicle will stop and remain at the postion as per my understanding. When the compacter reaches the park position it will no longer call this function.
After a timeout of 3 sec the AD vehicle will continue to drive and unload - implemented in AD.
When the AD vehicle leave the search range of the compacter it will continue it's work.