6.02.00045 Fixes #5802
We were using the presence of the isLeft attribute of wheels to figure
out if the wheel is a real wheel of an implement or just any support
wheel. We need the real wheels to determine the turning radius and the
distance between the attacher joint and the wheels for turn calculation.

Unfortunately the absence of isLeft isn't a reliable indicator of dummy
wheels anymore, so some implements, especially those with wheels on the
folding parts behave differently depending on their fold state when
starting the CP driver. This is possible a Giants bug.

Use hasTireTracks now instead of isLeft to find real wheels.