6.02.00060 Mode 2/3 improvements
Moving back before switching to unload course when close
to end of row.

Moving the safety-liveness balance towards liveness,
unloader yields to combine.

Move a bit faster than the combine while the pipe is opening
to keep the trailer in the pipe's range, so the combine won't
close it again while we are approaching.

Fixes #6065