6.02.00063 Turn fixes
Fixes #5855 #5598

If the headland is not perpendicular to the up/down rows we have
less space to turn on the field. Adjust headland width now if
the headland is at an angle.

For wide turns, if turn on field is activated now reverse back to
the field if the vehicle is off the field when it raises/deactivates
implements.