Pathfinder fix
The list of vehicles to check for collision was stored
in a global variable. This list is created every time the
pathfinder starts and contains all vehicles except the own
and the ones we want to ignore.

If a second pathfinder was started for another vehicle while
the first one was still running, the second vehicle's collision
data overwrote the first one's, for example including the first
vehicle in the list. So the first pathfinder was hitting bogus
collisions and found no path.

Made the collision list local to the pathfinder's context to
avoid overwriting it by another pathfinder.