6.03.00005 Fieldwork with front loader fixes #6412
Tractors with front loaders should now be working fine for fieldwork.

If there's no tool attached to the front loader the tractor may be
reversing instead of driving forward, in this case just move the front
loader up or down as it is in the way of the front proximity sensor.
Fixing this too would not be worth the effort.

improvements for mode 1, mode 3, mode 9

- implements mode 3 pipe positions
- implements mode 1 loading 2+ trailer/wilson with manual loading at start
- fixes mode 9 collision problems with trailers