Final fix for #6460 (mode 2/3)
The trailer heading was always 0, so when pathfinding started the
first trailer heading was always pointing to the East, regardless
of the actual heading of the trailer. This caused a bogus obstacle
detection when there was something (tree, other vehicle) to the
west of the trailer.

Initialize the trailer's heading on the start node before the
pathfinder starts.