Pathfinder Improvements (#6493)
* Final fix for #6460 (mode 2/3)

The trailer heading was always 0, so when pathfinding started the
first trailer heading was always pointing to the East, regardless
of the actual heading of the trailer. This caused a bogus obstacle
detection when there was something (tree, other vehicle) to the
west of the trailer.

Initialize the trailer's heading on the start node before the
pathfinder starts.

* Old pathfinder removed

* Pathfinder: ignore trailer when first checking the goal node

Since we don't know the trailer's heading at the goal
anyway.