6.03.00020 Mode2/3 improvements
- Wait before starting the next row if the
rendezvous point is close to the row end.
- use A* to calculate distance to combine before the
rendezvous.
- combine notifies unloader on missed rendezvous, unloader
re-plans route
- try to approach combine waiting after backing out of the fruit
from the rear so we won't cut in front of it.
- don't slow down around rendezvous when discharging already
- don't ask for a rendezvous when the combine is not willing to,
for example when unload is disabled on the first headland.
- won't initiate new rendezvous until the combine cancels the current one
- move up rendezvous points close to row end
- pipe in fruit map changed to have a 20 m buffer at each end of the
row to account for non-perpendicular headlands.
- fixed offsets when calculating target to combine, sometimes it
used the side offset as front offset.
- make sure to reset 95% full limit after making pocket
- added safety margin to the calculation of the distance until full
All this to avoid the situation where the combine stops before reaching
the rendezvous point because it thinks it is full (although only 95%)
- only considers unloaders actually waiting for assignment,
ignore the ones on unload course for example when a combine
is looking for an unloader
- if an unloader is already assigned to a combine, only it is
waiting for the combine to become ready for unload (for example
due to fruit in pipe), assign the same combine to it. This
should probably be refactored, adding a new state for the unloader,
like WAITING_FOR_COMBINE_TO_BECOME_READY or so.