No extra headland circle (#6605)
Fixed the calculation of the headland/middle transition.

If the center mode is lands, there is still a small chance that
an almost full circle connecting track will be added if the starting
point is very close to the start/end of the first working row of the
lands pattern.