6.03.00021 Course generator and turn fixes
No extra headland circle (#6605)

- Fixed the calculation of the headland/middle transition.

- If the center mode is lands, there is still a small chance that
an almost full circle connecting track will be added if the starting
point is very close to the start/end of the first working row of the
lands pattern.

Partial fix for #6383

- do not add straight section to turn end to align if there is no towed
implement
- if turn on field enabled but there's no space to turn (must reverse), use
the legacy wide turn (which is also buggy a bit).
- if you don't want your sprayers boom hit obstacles outside the field,
use a headland.