6.03.00024 Bridge and proximity sensor fixes
- Goal node elevation fixed (#6665)
- Proximity sensor directly on the back of the combine/chopper
allowing for the unloader to get closer.

When driving over non-level bridges, the controlled node isn't vertical and
the goal node was positioned on the ground level which is way below the vehicle.

Due to this, the goal's position in the controlled node's reference wasn't accurate
in this situation, seemed to be a lot closer when driving uphill, lot further when
driving downhill. Uphill goal being too close caused oscillations in PPC.

Moved the goal point on the level of the vehicle (controlled node), so the error
caused by the non-vertical controlled node is negligible.

Few proximity sensor and vehicle size fixes

- Proximity sensor sits on the back of the combine now
- fixed pathfinder vehicle data rectangles, clarified the
misconception that the direction node is always on the root node,
it is not, the direction node is the turning node, usually sits
between the non-steered wheels.