Mode2 (#6674)
* - devhelper shows proximity sensors
- off field penalty decreased

* - show proximity sensor in DevHelper

* - reduce off field penalty when combine on headland
- better driving time estimation to rendezvous point now
considering turns we have to make to get there

* Fix for #6633

Distance to next turn may not be defined for instance
in the last row.

* Off field penalty for first headland reduce

When unloading of the combine on the first headland is allowed,
reduce off-field penalty as we most likely will have to (and
since the user allowed it), will be able to drive off the field.

* Many pathfinder/Mode2 updates

- tried to make the rendezvous more fault tolerant
- relaxed pathfinder off-field constraints for the first two headlands
- use a higher fruit value limit in the hybrid parts as it is not a
problem getting temporarily closer to the fruit there, and it speeds
up the pathfinding quite a bit.

* don't show proximity sensor debug info

* small refactoring

* Mode 2/3 Improvements

- when the chopper/combine is turning, return the turn end wp
as the closest waypoint for the unloader to follow
(instead of the last passed one, which could be well before the
turn start). The logic in startFollowingChopper will decide
if it needs to start with the turn start or the turn end wp then.