6.03.00031 Fixes #6728, game freeze on load
Some modded vehicles (like the FS19_MerloP417) won't
correctly install themselves as vehicle types. This
prevents Courseplay from registering its event listeners
and thus, onLoad() is not called for this vehicle,
leaving an incomplete CP setup behind.

When AttacherJoints.attachImplement() is then called for
the vehicle during game load, it does have the Courseplay
specialization according to hasSpecialization() but it
is incomplete, leading to many errors.

So, check for vehicle.hasCourseplaySpec as well since that is
only set when the CP set up is complete.