Make overlap boxes parallel to the terrain
When looking for collisions with overlap boxes, the pathfinder was
setting the x and z rotation to 0, meaning the box was always
horizontal, even on sloped terrain.

This can lead to false positives if there are shapes close to the
terrain, therefore, make sure the overlap boxes are parallel to the
ground.