Mode 2 unload with AutoDrive (#7291)
Implementing feature request #5284

CP still needs a course for mode 2, but this can be any random 
one even on the opposite corner of the map, as this one isn't used at all.

Start CP driver with options 'Unloading' and use AutoDrive 
'refill/unload'. AutoDrive settings basically the same as for the 
fieldwork drivers: use pickup&deliver, select a proper field
exit strategy and take care about appropriate drive on
percentage (has to be lower than the setting in CP).