Jump Point Search integrated with game
Should work ok for fruit. There is no off-field penalty as
JPS works with a uniform cost grid only so any node is either
valid or invalid.

So for now, we consider nodes with fruit or obstacles invalid,
everything else is valid so do not expect to avoid your
neighbor's field ...