Fix offset course for multitool
When calculating the offset course we marked each waypoint on the
headland with lane = -1 for simplicity but that means that all
waypoints of the headland seem to be on the outermost headland.

Since the pathfinder turn drives on the outermost headland this
confused the algorithm completely and resulted in paths going
the wrong way around the field.

Change offset calculation so offset course keeps the original
lane number.